persistent-metrics-service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
apscheduler
httpx
pyyaml
pydantic
pydantic-settings
